requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.1
